Add files via upload
Added a method to check every rotation of a given string to see if it perfect in or out shuffles back to itself. 
Added a method to check every rotation of a given string to see if it perfect in or out shuffles then rotates back to itself. (not as useless as you expect)