data dump
Limited each file to 1 MB, as few will likely ever be used anyway. Because sequences vary in length, there will be a variable number of sequences in each file. The reason some are missing from 2-2 through 9-9 is because my program hit a stack overflow. As did the other sequence generator I've got, so I'm not just an asshole.